+++ src/components/Badge/Badge.styled.tsx
@@ -10,11 +10,9 @@ export const BadgeStyled = styled.mark`
     font-size           : 92%;
     text-align          : center;
     text-decoration     : none;
-    line-height         : 1;
     transition          : all 0.2s ease-in-out;
     background-position : center;
     user-select         : none;
-    padding             : 2px 4px;
     width               : fit-content;
     height              : fit-content;
 
@@ -42,27 +40,27 @@ export const BadgeStyled = styled.mark`
 
     //  DIFFERENT SIZES  //////////////////////////////////////////////////////
     &.size-tiny {
-        padding   : 4px 8px;
+        padding   : 2px 4px;
         font-size : ${(props : BadgeProps) => props.theme.text.paras.size * Math.pow(props.theme.text.headings.multiplier, -2)}em;
     }
 
     &.size-small {
-        padding   : 6px 12px;
+        padding   : 3px 6px;
         font-size : ${(props : BadgeProps) => props.theme.text.paras.size * Math.pow(props.theme.text.headings.multiplier, -1)}em;
     }
 
     &.size-medium {
-        padding   : 12px 24px;
+        padding   : 4px 8px;
         font-size : ${(props : BadgeProps) => props.theme.text.paras.size * Math.pow(props.theme.text.headings.multiplier, 1)}em;
     }
 
     &.size-large {
-        padding   : 16px 32px;
+        padding   : 6px 12px;
         font-size : ${(props : BadgeProps) => props.theme.text.paras.size * Math.pow(props.theme.text.headings.multiplier, 2)}em;
     }
 
     &.size-huge {
-        padding   : 24px 40px;
+        padding   : 8px 16px;
         font-size : ${(props : BadgeProps) => props.theme.text.paras.size * Math.pow(props.theme.text.headings.multiplier, 4)}em;
     }
 `;